Translate text

The translate function returns the translation of a word or phrase stored in the database or translation dataframe. Translations are identified by their unique ID.

tr(textid, language, table, id_col = 1L, con, tbl)

Arguments

textid
the unique ID of the string to translate

language
the language code to translate to. Users may specify a second language as a fallback by passing a vector of length 2.

table
a data.frame object (or something that can be coerced to one) with translations (optional, see details)

id_col
index or name of the column containing text IDs

con
a database connection containing necessary translation tables

tbl
the name of the table in the database containing the translations

Details

You can set the default language for your session using options(translatr.language = 'mi'), otherwise the first non-ID column in the translation table will be used. Similarly, you can specify the translation dictionary with the option translatr.table, or by supplying translatr.con and translatr.tbl to connect to a database with a table containing the translations.

The language codes do not necessarily need to follow any rule, since nothing is automated in the process - so long as your interface provides the correct language identifier, it will work. That said, we recommend you choose and stick to a standard format, such as those listed here https://en.wikipedia.org/wiki/List_of_ISO_639-1_codes. In the examples given, we use ISO 639-1.

Examples

trans_df <- data.frame(id = "hello", en = "Hello", mi = "Kia ora")
tr("hello", "mi", trans_df)
#> [1] "Kia ora"

require(RSQLite)
#> Loading required package: RSQLite
con <- dbConnect(SQLite(), system.file("introduction.sqlite", package = "translatr"))
options(translatr.con = con, translatr.tbl = "intro")
tr("hello", "fr")
#> [1] "Bonjour"
tr("statistics", "mi")
#> [1] "tatauranga"

# but there's no clever automatic translation going on:
DBI::dbReadTable(con, "intro")
#>           id         en         mi               fr
#> 1    welcome    Welcome  Haere mai        Bienvenue
#> 2      hello      Hello    Kia ora          Bonjour
#> 3 statistics statistics tatauranga les statistiques